requirements.txt
asgiref==3.8.1
channels==4.2.0
channels_redis==4.2.1
Django==5.1.3
django-otp==1.5.4
django-rest-passwordreset==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
msgpack==1.1.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.10.0
pytest==8.3.3
pytest-django==4.9.0
redis==5.2.0
sqlparse==0.5.2
